+++ CHANGELOG.md
@@ -61,7 +61,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 - `dispatch` module for [Dispatch](https://developer.apple.com/documentation/dispatch)
   library.
 
-  - Types: `DispatchObject`, `DispatchTime`.
+  - Types: `DispatchObject`, `DispatchTime`, `DispatchQos`, `DispatchQosClass`.
 
 - Pointer methods for `SEL`.